requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
coverage==5.3.1
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-autoslug==1.9.8
django-crispy-forms==1.10.0
django-localflavor==3.0.1
django-debug-toolbar==3.2
django-localflavor-br==1.0
django-model-utils==4.1.1
factory-boy==3.1.0
Faker==5.4.0
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
oauthlib==3.1.0
packaging==20.8
parso==0.8.1
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.9
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
Pygments==2.7.3
PyJWT==2.0.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
style==1.1.0
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
update==0.0.1
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
django-widget-tweaks==1.4.8
pycorreios3~=1.2.2
